+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.1] - 2020-04-27
+
 ### Added
 
 - `ParseRedisURI` helper function is added to create a `RedisConnOpt` from a URI string.